Move projects/listItem -> projects/table
