+++ src/js/components/projects/table/collaboratorCell.tsx
@@ -1,0 +1,26 @@
+import DataTableCell from '@salesforce/design-system-react/components/data-table/cell';
+import classNames from 'classnames';
+import React from 'react';
+
+import { TableCellProps } from '@/components/projects/table';
+
+const CollaboratorTableCell = ({
+  className,
+  children,
+  ...props
+}: TableCellProps & { children?: number }) => (
+  <DataTableCell
+    {...props}
+    className={classNames(
+      className,
+      'project-collaborators-cell',
+      'numbers-cell',
+      'complex-cell',
+    )}
+  >
+    {children}
+  </DataTableCell>
+);
+CollaboratorTableCell.displayName = DataTableCell.displayName;
+
+export default CollaboratorTableCell;
